Struct peel_ip::layer1::arp::ArpPacket

pub struct ArpPacket {
    pub hardware_type: ArpHardwareType,
    pub protocol_type: EtherType,
    pub hardware_length: u8,
    pub protocol_length: u8,
    pub operation: ArpOperation,
    pub sender_hardware_address: MacAddress,
    pub sender_protocol_address: Ipv4Addr,
    pub target_hardware_address: MacAddress,
    pub target_protocol_address: Ipv4Addr,
}

Representation of the Arp structure

Fields
hardware_type: ArpHardwareType

This field specifies the network protocol type. Example: Ethernet is 1.

protocol_type: EtherType

This field specifies the internetwork protocol for which the ARP request is intended. For IPv4, this has the value 0x0800. The permitted PTYPE values share a numbering space with those for EtherType.

hardware_length: u8

Length (in octets) of a hardware address. Ethernet addresses size is 6.

protocol_length: u8

Length (in octets) of addresses used in the upper layer protocol. (The upper layer protocol specified in PTYPE.) IPv4 address size is 4.

operation: ArpOperation

Specifies the operation that the sender is performing: 1 for request, 2 for reply.

sender_hardware_address: MacAddress

Media address of the sender.

sender_protocol_address: Ipv4Addr

Internetwork address of the sender.

target_hardware_address: MacAddress

Media address of the intended receiver.

target_protocol_address: Ipv4Addr

Target protocol address: Internetwork address of the intended receiver.