Enum peel_ip::layer4::http::HttpRequestMethod

pub enum HttpRequestMethod {
    Get,
    Post,
    Head,
    Put,
    Delete,
    Trace,
    Options,
    Connect,
    Patch,
}

List of supported HTTP request methods

Variants
Get

The GET method requests a representation of the specified resource.

Post

The POST method requests that the server accept the entity enclosed in the request as a new subordinate of the web resource identified by the URI.

Head

The HEAD method asks for a response identical to that of a GET request, but without the response body.

Put

The PUT method requests that the enclosed entity be stored under the supplied URI.

Delete

The DELETE method deletes the specified resource.

Trace

The TRACE method echoes the received request so that a client can see what (if any) changes or additions have been made by intermediate servers.

Options

The OPTIONS method returns the HTTP methods that the server supports for the specified URL.

Connect

The CONNECT method converts the request connection to a transparent TCP/IP tunnel, usually to facilitate SSL-encrypted communication (HTTPS) through an unencrypted HTTP proxy.

Patch

The PATCH method applies partial modifications to a resource.