Enum peel_ip::prelude::IpAddr

pub enum IpAddr {
    V4(Ipv4Addr),
    V6(Ipv6Addr),
}

An IP address, either an IPv4 or IPv6 address.

Examples

Constructing an IPv4 address:

use std::net::{IpAddr, Ipv4Addr};

IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));

Constructing an IPv6 address:

use std::net::{IpAddr, Ipv6Addr};

IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 1));

Variants
V4(Ipv4Addr)

Representation of an IPv4 address.

V6(Ipv6Addr)

Representation of an IPv6 address.

Methods

impl IpAddr

fn is_unspecified(&self) -> bool

Returns true for the special 'unspecified' address (IPv4, IPv6).

Examples

use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};

assert_eq!(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)).is_unspecified(), true);
assert_eq!(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 0)).is_unspecified(), true);

fn is_loopback(&self) -> bool

Returns true if this is a loopback address (IPv4, IPv6).

Examples

use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};

assert_eq!(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)).is_loopback(), true);
assert_eq!(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 0x1)).is_loopback(), true);

fn is_global(&self) -> bool

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns true if the address appears to be globally routable (IPv4, IPv6).

Examples

#![feature(ip)]

use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};

fn main() {
    assert_eq!(IpAddr::V4(Ipv4Addr::new(80, 9, 12, 3)).is_global(), true);
    assert_eq!(IpAddr::V6(Ipv6Addr::new(0, 0, 0x1c9, 0, 0, 0xafc8, 0, 0x1)).is_global(),
               true);
}

fn is_multicast(&self) -> bool

Returns true if this is a multicast address (IPv4, IPv6).

Examples

use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};

assert_eq!(IpAddr::V4(Ipv4Addr::new(224, 254, 0, 0)).is_multicast(), true);
assert_eq!(IpAddr::V6(Ipv6Addr::new(0xff00, 0, 0, 0, 0, 0, 0, 0)).is_multicast(), true);

fn is_documentation(&self) -> bool

Unstable (ip)
: extra functionality has not been scrutinized to the level that it should be stable

Returns true if this address is in a range designated for documentation (IPv4, IPv6).

Examples

#![feature(ip)]

use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};

fn main() {
    assert_eq!(IpAddr::V4(Ipv4Addr::new(203, 0, 113, 6)).is_documentation(), true);
    assert_eq!(IpAddr::V6(Ipv6Addr::new(0x2001, 0xdb8, 0, 0, 0, 0, 0, 0))
                      .is_documentation(), true);
}

fn is_ipv4(&self) -> bool

Unstable (ipaddr_checker)

Returns true if this address is a valid IPv4 address, false if it's a valid IPv6 address.

Examples

#![feature(ipaddr_checker)]

use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};

fn main() {
    assert_eq!(IpAddr::V4(Ipv4Addr::new(203, 0, 113, 6)).is_ipv4(), true);
    assert_eq!(IpAddr::V6(Ipv6Addr::new(0x2001, 0xdb8, 0, 0, 0, 0, 0, 0)).is_ipv4(),
               false);
}

fn is_ipv6(&self) -> bool

Unstable (ipaddr_checker)

Returns true if this address is a valid IPv6 address, false if it's a valid IPv4 address.

Examples

#![feature(ipaddr_checker)]

use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};

fn main() {
    assert_eq!(IpAddr::V4(Ipv4Addr::new(203, 0, 113, 6)).is_ipv6(), false);
    assert_eq!(IpAddr::V6(Ipv6Addr::new(0x2001, 0xdb8, 0, 0, 0, 0, 0, 0)).is_ipv6(),
               true);
}