Struct peel_ip::prelude::str::CharIndices

pub struct CharIndices<'a> { /* fields omitted */ }

Iterator for a string's characters and their byte offsets.

Methods

impl<'a> CharIndices<'a>

fn as_str(&self) -> &'a str

View the underlying data as a subslice of the original data.

This has the same lifetime as the original slice, and so the iterator can continue to be used while this exists.